+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/country_specific/germany/de_tax_number_recognizer.py
@@ -34,6 +34,8 @@ class DeTaxNumberRecognizer(PatternRecognizer):
     :param supported_entity: The entity this recognizer can detect
     """
 
+    COUNTRY_CODE = "de"
+
     PATTERNS = [
         Pattern(
             "Steuernummer ELSTER (bundeseinheitlich, 13-stellig)",