+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/country_specific/india/in_aadhaar_recognizer.py
@@ -18,6 +18,8 @@ class InAadhaarRecognizer(PatternRecognizer):
     This can allow a greater variety in input, for example by removing dashes or spaces.
     """
 
+    COUNTRY_CODE = "in"
+
     PATTERNS = [
         Pattern(
             "AADHAAR (Very Weak)",