+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/country_specific/korea/kr_brn_recognizer.py
@@ -25,6 +25,8 @@ class KrBrnRecognizer(PatternRecognizer):
     to be used during pattern matching (e.g., removing dashes).
     """
 
+    COUNTRY_CODE = "kr"
+
     PATTERNS = [
         Pattern(
             "BRN (Weak)",