+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/country_specific/sweden/se_organisationsnummer_recognizer.py
@@ -17,6 +17,8 @@ class SeOrganisationsnummerRecognizer(PatternRecognizer):
     * Luhn checksum validation
     """
 
+    COUNTRY_CODE = "se"
+
     PATTERNS = [
         Pattern(
             "Swedish Organisationsnummer (Medium)",