+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/country_specific/uk/uk_driving_licence_recognizer.py
@@ -29,6 +29,8 @@ class UkDrivingLicenceRecognizer(PatternRecognizer):
     :param supported_entity: The entity this recognizer can detect
     """
 
+    COUNTRY_CODE = "uk"
+
     PATTERNS = [
         Pattern(
             "UK Driving Licence",