+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/country_specific/us/us_itin_recognizer.py
@@ -13,6 +13,8 @@ class UsItinRecognizer(PatternRecognizer):
     :param supported_entity: The entity this recognizer can detect
     """
 
+    COUNTRY_CODE = "us"
+
     PATTERNS = [
         Pattern(
             "Itin (very weak)",